+++ spatialprofilingtoolbox/workflow/halo_import/__init__.py
@@ -1,11 +1,10 @@
 """The main data import workflow."""
 from spatialprofilingtoolbox.workflow.defaults.workflow_module_exporting import WorkflowModules
 
-from \
-    spatialprofilingtoolbox.workflow.dataset_designs.multiplexed_imaging.halo_cell_metadata_design \
-    import HALOCellMetadataDesign
+from spatialprofilingtoolbox.workflow.dataset_designs.multiplexed_imaging.\
+    halo_cell_metadata_design import HALOCellMetadataDesign
 
-from spatialprofilingtoolbox.workflow.defaults.job_generator import JobGenerator
+from spatialprofilingtoolbox.workflow.halo_import.job_generator import JobGenerator
 from spatialprofilingtoolbox.workflow.halo_import.initializer import HALOImportInitializer
 from spatialprofilingtoolbox.workflow.halo_import.computational_design import HALOImportDesign
 from spatialprofilingtoolbox.workflow.halo_import.core import HALOImportCoreJob
