+++ spatialprofilingtoolbox/workflow/halo_import/initializer.py
@@ -1,32 +1,31 @@
 """The initializer of the main data import workflow."""
-from typing import Optional
+from abc import ABC, abstractmethod
 
 from spatialprofilingtoolbox.workflow.defaults.cli_arguments import add_argument
-from spatialprofilingtoolbox.workflow.defaults.initializer import Initializer
 from spatialprofilingtoolbox.workflow.source_file_adi_parsing.skimmer import DataSkimmer
 from spatialprofilingtoolbox.standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
 
 
+class Initializer(ABC):
+    """Interface for the intializer job for the Nextflow-managed workflows."""
+    def __init__(self, dataset_design=None, computational_design=None, **kwargs): # pylint: disable=unused-argument
+        self.dataset_design = dataset_design
+        self.computational_design = computational_design
+
+    @staticmethod
+    @abstractmethod
+    def solicit_cli_arguments(parser):
+        pass
+
+    @abstractmethod
+    def initialize(self, **kwargs):
+        pass
+
+
 class HALOImportInitializer(Initializer):
     """Initial process for main data import workflow; does most of the import."""
-    def __init__(self,
-                 file_manifest_file=None,
-                 elementary_phenotypes_file=None,
-                 composite_phenotypes_file=None,
-                 outcomes_file=None,
-                 compartments_file=None,
-                 subjects_file=None,
-                 **kwargs,
-                 ):
-        super().__init__(**kwargs)
-        self.file_manifest_file = file_manifest_file
-        self.elementary_phenotypes_file = elementary_phenotypes_file
-        self.composite_phenotypes_file = composite_phenotypes_file
-        self.outcomes_file = outcomes_file
-        self.compartments_file = compartments_file
-        self.subjects_file = subjects_file
 
     @staticmethod
     def solicit_cli_arguments(parser):
@@ -41,12 +40,8 @@ class HALOImportInitializer(Initializer):
         add_argument(parser, 'diagnosis file')
         add_argument(parser, 'interventions file')
 
-    def initialize(
-        self,
-        database_config_file: Optional[str] = None,
-        **kwargs,
-    ):
-        if database_config_file is None:
+    def initialize(self, database_config_file: str = '', **kwargs):
+        if database_config_file == '':
             message = 'Need to supply database configuration file.'
             logger.error(message)
             raise ValueError(message)
