+++ spatialprofilingtoolbox/workflow/phenotype_proximity/initializer.py
@@ -1,5 +1,5 @@
 """The initializer for the phenotype proximity workflow."""
-from spatialprofilingtoolbox.workflow.defaults.initializer import Initializer
+from spatialprofilingtoolbox.workflow.halo_import.initializer import Initializer
 from spatialprofilingtoolbox.workflow.defaults.cli_arguments import add_argument
