+++ spatialprofilingtoolbox/workflow/scripts/aggregate_core_results.py
@@ -33,8 +33,6 @@ if __name__ == '__main__':
     from spatialprofilingtoolbox.standalone_utilities.module_load_error import \
         SuggestExtrasException
     try:
-        from spatialprofilingtoolbox.workflow.defaults.computational_design import \
-            ComputationalDesign
         from \
             spatialprofilingtoolbox.workflow.dataset_designs.multiplexed_imaging.\
                 halo_cell_metadata_design import HALOCellMetadataDesign
@@ -45,8 +43,7 @@ if __name__ == '__main__':
         w.computational_design
         for w in workflows.values()
     ]
-    for module in computational_designs + [HALOCellMetadataDesign, ComputationalDesign]:
-        module.solicit_cli_arguments(parser)
+    HALOCellMetadataDesign.solicit_cli_arguments(parser)
 
     args = vars(parser.parse_args())
